- created sessionContext - created sessionPRovider and added to root
