+++ src/context/sessionManager.tsx
@@ -1,24 +1,7 @@
-import { createContext, useContext, useRef, useState } from "react";
+import { useRef, useState } from "react";
 import type { Task, Block } from "../types";
 import { useBlock } from "./blockContext";
-
-interface SessionContextType {
-  activeTask: Task | null;
-  activeBlock: Block | null;
-  start: (block: Block, task: Task) => void;
-  pause: () => void;
-  reset: () => void;
-}
-
-const SessionContext = createContext<SessionContextType>({
-  activeTask: null,
-  activeBlock: null,
-  start: () => {},
-  pause: () => {},
-  reset: () => {},
-});
-
-export const useSession = () => useContext(SessionContext);
+import { SessionContext } from "./sessionContext";
 
 export const SessionProvider = ({
   children,
@@ -31,12 +14,11 @@ export const SessionProvider = ({
   const [activeTask, setActiveTask] = useState<Task | null>(null);
   const [activeBlock, setActiveBlock] = useState<Block | null>(null);
 
-  // GLOBAL REFS (never lost across navigation)
+  // GLOBAL REFS
   const intervalRef = useRef<NodeJS.Timeout | null>(null);
   const elapsedRef = useRef<number>(0);
   const remainingRef = useRef<number>(0);
 
-  // THESE WILL BE IMPLEMENTED AS WE GO
   const start = (block: Block, task: Task) => {};
   const pause = () => {};
   const reset = () => {};
